requirements.txt
Django==2.2
django-cors-headers==3.1.0
django-filter==2.2.0
djangorestframework==3.10.3
coreapi==2.3.3
coreschema==0.0.4
MarkupSafe==1.0
Markdown==2.6.11
psycopg2-binary==2.8.3
flake8==3.7.8
gunicorn==19.9.0
django-heroku==0.3.1
